+++ src/StateSmithTest/TriggerModHelper_Tests.cs
@@ -11,6 +11,34 @@ using Xunit;
 
 namespace StateSmithTest;
 
+public class SupportAlternateTriggersProcessor_Tests
+{
+    // https://github.com/StateSmith/StateSmith/issues/108
+    [Fact]
+    public void TestEntry()
+    {
+        var inputSmBuilder = new InputSmBuilder();
+
+        inputSmBuilder.ConvertPlantUmlTextNodesToVertices("""
+            @startuml SomeSmName
+            [*] --> MY_STATE_1
+            MY_STATE_1 : enter / log("enter");
+            MY_STATE_1 : entry / log("entry");
+            @enduml
+        """);
+
+        inputSmBuilder.FinishRunning();
+        var myState1 = inputSmBuilder.GetStateMachine().Child<State>("MY_STATE_1");
+
+        var behaviorMatcher = VertexTestHelper.BuildFluentAssertionBehaviorMatcher(actionCode: true, guardCode: true, transitionTarget: true, triggers: true);
+
+        myState1.Behaviors.Should().BeEquivalentTo(new List<Behavior>() {
+            Behavior.NewEnterBehavior("""log("enter");"""),
+            Behavior.NewEnterBehavior("""log("entry");"""),
+        }, behaviorMatcher);
+    }
+}
+
 public class TriggerModHelper_Tests
 {
     [Fact]